# apps/backend/openapi.json
{
  "components": {
    "schemas": {
...
      "ApiErrorDto": {
        "properties": {
          "error": {
            "description": "Short error category",
            "example": "Bad Request",
            "type": "string"
          },
          "message": {
            "description": "Human-readable message or array of validation messages",
            "example": "validation failed",
            "oneOf": [
              {
                "type": "string"
              },
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              }
            ]
          },
          "path": {
            "description": "Request path",
            "example": "/api/v1/dashboard/bookings",
            "type": "string"
          },
          "requestId": {
            "description": "Correlation ID for log lookup",
            "example": "req-7f9c2e1a",
            "type": "string"
          },
          "statusCode": {
            "description": "HTTP status code",
            "example": 400,
            "type": "number"
          },
          "timestamp": {
            "description": "ISO timestamp",
            "example": "2026-04-17T12:34:56.000Z",
            "type": "string"
          }
        },
        "required": [
          "statusCode",
          "error",
          "message",
          "timestamp",
          "path"
        ],
        "type": "object"
      },
      "ApplyCouponDto": {
        "properties": {
          "clientId": {
            "description": "Client redeeming the coupon",
            "example": "00000000-0000-0000-0000-000000000000",
            "type": "string"
          },
          "code": {
            "description": "Coupon code to apply",
            "example": "WELCOME10",
            "type": "string"
          },
          "invoiceId": {
            "description": "Invoice to apply the coupon to",
            "example": "00000000-0000-0000-0000-000000000000",
            "type": "string"
          }
        },
        "required": [
          "invoiceId",
          "clientId",
          "code"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {
      "email": "dev@carekit.dev",
      "name": "CareKit Engineering",
      "url": "https://carekit.dev"
    },
    "description": "CareKit clinic management platform — dashboard & mobile API",
    "license": {
      "name": "Proprietary",
      "url": "https://carekit.dev/license"
    },
    "title": "CareKit API",
    "version": "2.0"
  },
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/dashboard/finance/coupons": {
      "get": {
        "operationId": "DashboardFinanceController_listCouponsEndpoint",
        "parameters": [
          {
            "description": "Search by coupon code or description",
            "in": "query",
            "name": "search",
            "required": false,
            "schema": {
              "example": "WELCOME",
              "type": "string"
            }
          },
          {
            "description": "Filter by coupon status",
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "enum": [
                "active",
                "inactive",
                "expired"
              ],
              "example": "active",
              "type": "string"
            }
          },
          {
            "description": "Page number (1-based)",
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "description": "Records per page (max 100)",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "example": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated coupon list"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "Validation failed"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "Missing or invalid authentication"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "Action denied by permission policy"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "Unhandled server error"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List coupons",
        "tags": [
          "Dashboard / Finance"
        ]
      },
      "post": {
        "operationId": "DashboardFinanceController_createCouponEndpoint",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCouponDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Coupon created"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "Validation failed"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "Missing or invalid authentication"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "Action denied by permission policy"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "Unhandled server error"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a coupon",
        "tags": [
          "Dashboard / Finance"
        ]
      }
    },
    "/api/v1/dashboard/finance/coupons/apply": {
      "post": {
        "operationId": "DashboardFinanceController_applyCouponEndpoint",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApplyCouponDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Coupon applied; returns updated discount amount"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "Validation failed"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "Missing or invalid authentication"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "Action denied by permission policy"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "Unhandled server error"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Apply a coupon code to an invoice",
        "tags": [
          "Dashboard / Finance"
        ]
      }
    },
    "/api/v1/dashboard/finance/coupons/{id}": {
      "delete": {
        "operationId": "DashboardFinanceController_deleteCouponEndpoint",
        "parameters": [
          {
            "description": "Coupon UUID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "example": "00000000-0000-0000-0000-000000000000",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Coupon deleted"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "Validation failed"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "Missing or invalid authentication"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "Action denied by permission policy"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "Coupon not found"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "Unhandled server error"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete a coupon",
        "tags": [
          "Dashboard / Finance"
        ]
      },
      "get": {
        "operationId": "DashboardFinanceController_getCouponEndpoint",
        "parameters": [
          {
            "description": "Coupon UUID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "example": "00000000-0000-0000-0000-000000000000",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Coupon found"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "Validation failed"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "Missing or invalid authentication"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "Action denied by permission policy"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "Coupon not found"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "Unhandled server error"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get a coupon by id",
        "tags": [
          "Dashboard / Finance"
        ]
      },
      "patch": {
        "operationId": "DashboardFinanceController_updateCouponEndpoint",
        "parameters": [
          {
            "description": "Coupon UUID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "example": "00000000-0000-0000-0000-000000000000",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCouponDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Coupon updated"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "Validation failed"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "Missing or invalid authentication"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "Action denied by permission policy"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "Coupon not found"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "Unhandled server error"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update a coupon",
        "tags": [
          "Dashboard / Finance"
        ]
      }
    },
    "/api/v1/dashboard/finance/invoices": {
      "post": {
        "operationId": "DashboardFinanceController_createInv",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateInvoiceDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Invoice created"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "Validation failed"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "Missing or invalid authentication"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "Action denied by permission policy"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "Unhandled server error"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create an invoice",
        "tags": [
          "Dashboard / Finance"
        ]
      }
    },
    "/api/v1/dashboard/finance/invoices/{id}": {
      "get": {
        "operationId": "DashboardFinanceController_getInv",
        "parameters": [
          {
            "description": "Invoice UUID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "example": "00000000-0000-0000-0000-000000000000",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Invoice found"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "Validation failed"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "Missing or invalid authentication"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "Action denied by permission policy"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "Invoice not found"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "Unhandled server error"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get an invoice by id",
        "tags": [
          "Dashboard / Finance"
        ]
      }
    },
...
    "/api/v1/dashboard/finance/payments/bank-transfer": {
      "post": {
        "operationId": "DashboardFinanceController_bankTransferEndpoint",
        "parameters": [],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "properties": {
                  "amount": {
                    "example": 100,
                    "type": "number"
                  },
                  "clientId": {
                    "example": "00000000-0000-0000-0000-000000000000",
                    "format": "uuid",
                    "type": "string"
                  },
                  "invoiceId": {
                    "example": "00000000-0000-0000-0000-000000000000",
                    "format": "uuid",
                    "type": "string"
                  },
                  "receipt": {
                    "description": "Bank transfer receipt image or PDF",
                    "format": "binary",
                    "type": "string"
                  }
                },
                "required": [
                  "receipt",
                  "invoiceId",
                  "clientId",
                  "amount"
                ],
                "type": "object"
              }
            }
          },
          "description": "Receipt file + invoice metadata",
          "required": true
        },
        "responses": {
          "201": {
            "description": "Bank transfer receipt uploaded"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "Validation failed"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "Missing or invalid authentication"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "Action denied by permission policy"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "Unhandled server error"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Upload a bank transfer receipt for an invoice",
        "tags": [
          "Dashboard / Finance"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "Local dev",
      "url": "http://localhost:5100"
    }
  ],
...
}